bigslice: distributed data processing on top of bigmachine
Summary:
Package bigslice implements a distributed data processing system.
Users compose computations by operating over large collections ("big
slices") of data, transforming them with a handful of combinators.
While users express computations using collections-style operations,
bigslice takes care of the details of parallel execution and
distribution across multiple machines.

Bigslice jobs can run locally, but uses bigmachine for distribution
among a cluster of compute nodes. In either case, user code does not
change; the detials of distribution are handled by the combination
of bigmachine and bigslice.

Because Go cannot easily serialize code to be sent over the wire and
executed remotely, bigslice programs have to be written with a few
constraints:

 1. All slices must be constructed by bigslice funcs (bigslice.Func),
    and all such functions must be instantiated before
    bigslice.Start is called. This rule is easy to follow: if funcs
    are global variables, and bigslice.Start is called from a
    program's main, then the program is compliant.
 2. The driver program must be compiled on the same GOOS and GOARCH
    as the target architecture. When running locally, this is not a
    concern, but programs that require distribution must be run from
    a linux/amd64 binary.

Note that this change was whittled down in order to attempt to
include just the "kernel" of bigmachine. It should be considered
"minimally viable", though it is already useful: for example, I have
ported methylation kmer counts baseline building to using this system
without any issues.

It is missing several important features, including:

 1. Fault tolerance for bigmachines;
 2. Disk spilling of task outputs;
 3. Better main/tooling/opinion on how to invoke bigslice programs.

However none of these change the essence, structure, or design of the
system as is.

Reviewers: joshnewman, pgopal, jheart

Reviewed By: pgopal

Subscribers: cnicolaou

Differential Revision: https://phabricator.grailbio.com/D11896

fbshipit-source-id: 75a899a