bigslice: store task output to disk when using bigmachine
Summary:
This change uses an on-disk encoding to store task output. The output
is stored as gob-encoded batches of rows in column-major order. Task
output follows the ideas from the Monotasks[1] paper: rows are first
serialized in  memory while performing the (streaming) compute task.
When this is done, it is written to disk. The two phases aren't yet
scheduled separately: this will be a follow-up change.

Along the way we simplify the Reader interface: implementations now
no longer need to consider AllPartitions. This is never useful except
for diagnostics and complicates Reader implementations, especially in
the face of serialized storage.

We also use the newly added bigmachine testsystem to test the
bigmachine executor.

[1] http://kayousterhout.org/publications/sosp17-final183.pdf

Reviewers: pgopal, joshnewman, jheart

Reviewed By: jheart

Differential Revision: https://phabricator.grailbio.com/D12303

fbshipit-source-id: d439fd4