bigslice: stream task output directly to the configured store
Summary:
Previously we buffered task output in memory, with the idea of using
monotask scheduling on each local node. This requires a little more
work to be feasible: specifically, we need to manage (buffer) memory
as a resource, or else we can run out of memory buffering task
output.

For now, stream data directly.

Reviewers: pgopal, jheart, joshnewman

Reviewed By: jheart

Differential Revision: https://phabricator.grailbio.com/D12339

fbshipit-source-id: 73e5b54