bigslice: refactor []reflect.Value to Frame
Summary:
Much of the implemention of slice ops and evaluation is made complex
by the fact that values are stored, transmitted, and operated on in a
column-oriented fashion. In particular, this caused most data to be
passed around as []reflect.Value, which was really a slice of column
vectors (themselves slices). This change attempts to clear up and
simplify the code by introducing a lightweight Frame abstraction: a
Frame is a []reflect.Value, representing a list of column vectors.
Frame defines a number of common operations. The result is that the
code reads much closer to its intent as we can uniformly operate on
columns (via indexing Frames) and rows (via methods on Frames).

We also introduce a Type interface, which describes the type of a
Frame (and Slice, and Task), yielding uniform handling of type
checking and frame creation.

Reviewers: pgopal, jheart

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D13056

fbshipit-source-id: 5f1ddaf