bigslice/mapio: sorted, on-disk map
Summary:
Package mapio implements a sorted, on-disk map, similar to the
SSTable data structure used in Bigtable, Cassandra, and others. Maps
are read-only, and are produced by a Writer. Each Writer expects keys
to be appended in lexicographic order. Buf provides a means of
buffering writes to be sorted before appended to a Writer.

Mapio will be used in bigslice to implement joins (it can support
both merge-sort as well as hash-joins) and also fold/reduce
operations where state is too big to fit in memory (currently these
are implemented as in-memory hash joins).

Mapios can be efficiently merged. The implementation includes an
in-memory reorder buffer that can be used to implement external
sorting of large datasets.

Mapio is likely useful outside of bigslice: in particular, a number
of uses of recordio that use simple key indices could make use of
mapios instead.

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito, pgopal, jheart

Differential Revision: https://phabricator.grailbio.com/D13184

fbshipit-source-id: 0182c96