bigslice: allow user-defined custom hash functions
Summary:
This change uses user-defined hash functions for types that implement

	Hash32() uint32

Note that this operation currently requires an interface conversion
for each row, which is not ideal. It may be possible to perform
method invocation in an unsafe manner since we can guarantee that all
the elements of the slice are of the same type.

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D13398

fbshipit-source-id: aa7a45e