bigslice: separate invocations from task execution
Summary:
Bigslice executors are called with a task and the invocation
from which the task was compiled. This allowed a remote
process to reproduce the task graph and name individual
tasks within it.

There are a few downsides to this approach:

(1) If argument values are large (e.g., parameters for a machine
learning task), then we're needlessly re-transmitting these for each
task executed.
(2) It imposes a task reuse model: namely, that each unique
invocation produces only one set of tasks. This should instead be
managed by the evaluator so that we can execute tasks entirely for
their side effects.
(3) We cannot compose function invocations: for example, a task
from one invocation cannot reference one from another.

This change separates function invocation and slice compilation
from task execution. Specifically:

(1) Workers are instructed to invoke a function and compile its slice
explicitly by the evaluator. The invocation (i.e., its compiled task
graph) is named by the evaluator, using a evaluator-global index.
(2) The evaluator keeps track of which machines have performed which
invocations, and then automatically performs invocations when needed.
(3) Tasks are now named by their invocation index and task name.

With this change, we can now use a session to evaluate slices for
their side effects since each sess.Run produces a new invocation,
and we'll now transmit function parameters only once. (Occasionally
more times if there are invocation failures.)

It also lays the foundation for permitting tasks to cross invocation
boundaries, so that we can, for example, reuse task output without
having to explicitly output data to intermediate files.

Reviewers: pgopal, jheart

Reviewed By: jheart

Differential Revision: https://phabricator.grailbio.com/D13596

fbshipit-source-id: 139fda2