bigslice: make tasks stateful and the evaluator stateless
Summary:
This change adds explicit states to tasks, and also pushes
parallelism control and policy into the executors. After this change,
evaluators do not maintain task states, and we support concurrent
evaluations (i.e., a driver can run goroutines to manage multiple
evaluations concurrently against the same session).

The executors are responsible for managing task states, and the Task
struct gains a number of facilities for doing so.

This change lays the groundwork for proper fault tolerance: after
this change, the executors can report lost tasks and the evaluators
can restart them (on different machines) when needed; the evaluator
need only restart evaluation to handle such faults.

The change also allows us to push parallelism policies into the
executors, and the bigmachine executor is now free to allocate (and
relinquish) resources following demand.

Reviewers: pgopal, jheart

Reviewed By: jheart

Differential Revision: https://phabricator.grailbio.com/D13698

fbshipit-source-id: 9066f72