bigslice: introduce machineReader to lazily read streams from Machines
Summary:
Data are transmitted between nodes as raw bytestreams. This allows
Bigslice to store the bytestreams pre-encoded, and simply transmit
them when they are needed. Bigslice thus uses the Bigmachine
streaming RPC facility which in turn is just an API for HTTP/2
streams. These streams do implement backpressure, but we can buffer
too much data when concurrency levels are high.

This change introduces machineReader, which begins streaming
lazily--that is, upon the first invocation of Reader.Read. Thus when
we read data sequentially from a number of shards (for a single
partition of a slice), we don't end up buffering the initial set of
bytes for each shard, but rather transmit purely one at a time.

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D13783

fbshipit-source-id: fe9dd98