bigslice: shuffle partition shard read order
Summary:
This change shuffles the shard order in which reader tasks read a
partition. The idea is to avoid situations in which "thundering herd"
issues can occur, e.g., a large set of tasks all read shard 0 of
partitions 0-n, then shard 1, then shard 2, and so on.

We solve this problem by randomizing the order in which these
shards are read; eventually we may want to pursue proper load
balancing (though this can introduce problems of its own).

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D13785

fbshipit-source-id: 892786f