bigslice: first-class combiners, per-machine combine buffers
Summary:
This changes the way Bigslice implements combiners. They are now a
first-class property of Slices (and Tasks); when they are run, the
evaluator is responsible for combinging multiple outputs into a
single per-machine combine buffer.

This significantly improves the performance of map-side combining.
For example, in the "frequent fragments" GQL query, we get key
utilization (average hits) of 10 while maintaining a relatively small
memory footprint.

In this implementation, we maintain an in-memory combine buffer that
spills to disk after a threshold has been reached. This setup
complicates fault tolerance since each machine maintains a single
output for multiple tasks, but only a little: a future change will
move the responsibility of assigning combine keys to the evaluator,
at which point lost tasks need simply retrieve a new combine key, and
on read time, we need to compute a union that completes the set of
shards for a given combine key.

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D14259

fbshipit-source-id: 9938889