bigslice: add support for iterative slice computation
Summary:
Iterative computation is useful in machine learning and other
applications. This change permits results of a function application
(a Slice) to be passed to another. The result of a Func application
is now an exec.Result (which itself implements bigslice.Slice);
Results may be passed to other bigslice.Funcs expecting
bigslice.Slice arguments. Behind the scenes, these arguments are
resolved to the underlying invocation responsible for the result and
then reconstituted at compilation time on the worker side (in the
case of the bigmachine executor). Since this now permits, in
principle, arbitrary compute graphs to be constructed, the executor
must keep track of dependencies between invocations and then compile
the full graph as necessary.

Reviewers: pgopal, jheart

Reviewed By: pgopal

Subscribers: afields

Differential Revision: https://phabricator.grailbio.com/D14521

fbshipit-source-id: 58386a5