bigslice: improve partition-combine performance when there are hot keys (or few partitions)
Summary:
The current partition-combine operation maintains per-machine combine
buffers that require synchornization when combining values from
individual tasks. When key distribution is even, this does not
present any synchronization overhead as each partition is equally
likely to be combined: as long as #partitions >> #tasks (on a single
machine), this works out well. (This is usually the case.) However,
when there are hot keys or hot partitions, these become contended and
present a large synchronization overhead.

This change fixes this situation by adding per-task combine buffers.
Values are combined into these buffers without any synchronization,
and when they reach half of their capacity, the task attempts to
combine its least frequent keys into the per-machine combine buffer.
(In the extreme case where there is a single key, this means that all
combines will be performed entirely locally, and synchronization is
required only to combine these values again when the tasks are
completed.)

Reviewers: ysaito, pgopal, jheart

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D14968

fbshipit-source-id: eb4bf86