grailfile: localFile.CloseNoSync
Summary:
Add a public method, CloseNoSync, to the local file implementation to
avoid syncing while closing a file. Use this to close files in
Bigslice. There is some evidence that this causes sync pileups on EBS
in Bigmachine/Bigslice. In this case, the lifetime of the disk is
anyway tied to the machine, so there's no need to sync.

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D14954

fbshipit-source-id: f5556d1