bigslice/frame: define frame.Column as reflect.Value
Summary:
This change does not alter any functionality: it defines frame as

	type Column reflect.Value
	type Frame []Column

In order to clarify the various uses of reflect.Value in the code
base. Columns now require explicit conversion to Values: these are
exactly the points in the code base that warrants extra attention.

It also makes intents clearer by using a more appropriate name for
column values.

Reviewers: pgopal, jheart

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D15126

fbshipit-source-id: ffc18e3