bigslice: new, efficient frame implementation
Summary:
This change introduces a new frame implementation that avoids most
heap allocations in the hot paths. It does this by two primary means:

- We maintain our own slice headers, and share underlying data by
keeping offsets. Frames retain the semantics of Go slices, but we can
efficiently operate on them without escaping intermediate values to
heap.

- Operations are precomputed and shared among every allocated slice,
and type checks are precomputed where they can be.

With this change, we avoid almost all hot-path allocation in
Bigslice, speeding up the "frequent fragments" query by nearly 2.5x
in my tests.

This change also lays the groundwork for more work on this front, and
gives us a clean Frame abstraction that can multiplex abstract other
physical layouts in the future (e.g., flatbuffers/feathers style
in-line data).

This change also introduces a new combiner implementation which
avoids another category of heap allocations. Instead of using
external indexers, we implement a hash table directly on top of a
Frame. This both simplifies the kernel operations needed (only Hash()
and Less() are needed) and at the same time reduces the memory and
allocation requirements inmposed by an external map: the data are
already there; the map was duplicative, and seemingly expensive in GC
work.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D15717

fbshipit-source-id: e80d5c0