Fix a bug in the bigslice sorter; it emits garbage keys.
Summary:
The problem is that it overwrites a frame while a key/value pair being emitted
is still refering to the old values in the frame. This issue was obscured
because all the manipulations happen through reflections.

I also added a few debug logging statements to help tracing.

Bonus fix: in bigmachine:845, if pcomb is empty, it crashes topn later.

Reviewers: marius, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D15878

fbshipit-source-id: d19f5bb