bigslice/exec: limit the number of concurrent commits, and commit only from the evaluator
Summary:
Previously, each worker attempted to commit combiners for each task
dependency. This caused a lot of redundant RPCs and churn. Instead,
we guarantee that combiners are committed once a task is submitted to
the executor.

Second, we limit the number of concurrent commits that are allowable.
This caused a lot of thrashing as we were performing
on large jobs.

With this change, I can run a large GQL query without any
unresponsive hosts; keepalive reply times never exceed a couple of
seconds, and no RPC client resets were observed.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D16132

fbshipit-source-id: 2a5d546