bigslice, bigslice/exec: add ability to declare machine-exclusive funcs
Summary:
This introduces FuncValue.Exclusive() which marks a Func as
machine-exclusive, meaning that at most one invocation of this
function may be performed on any underlying machine.

The use case here is to allow jobs to partition memory: jobs may load
large datastructures into memory and this forces locality of tasks
within such a job. (Alex F. has one such case for scoring, where we
would benefit not only from locality, but also the fact that we can't
even fit more than one model into a single machine's memory at a
time.)

This is a rather coarse tool, of course, and we may provide a more
refined mechanism in the future (e.g., the ability to declare
"resources" or even provide expected resource footprints on funcs).
There is one more hitch with the current implementation: this is
implemented by running a separate cluster for each exclusive
invocation index; and thus the total parallelism is multiplied by the
number of invocations. I think this is ok for now until while we
figure out a more elegant mechanism.

Reviewers: ysaito, afields, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D17075

fbshipit-source-id: 68c53cf