bigslice/frame: fix implementation of frame.Value
Summary:
Frame.Value could return truncated frames when they were constructed
by frame.AppendFrame incrementally: when grown, the new internal
reflect.Value's length was the current length, not the allocated
length. This isn't a problem since we generally access the underlying
arrays directly with our own slice ops, but Cogroup accesses the
reflect.Value directly.

We also include a large-scale integration/regression test and a test
case for a large-scale cogroup. We can expand this binary with other
tests in the future.

Reviewers: pgopal, ysaito, cnicolaou

Reviewed By: ysaito, cnicolaou

Differential Revision: https://phabricator.grailbio.com/D17395

fbshipit-source-id: c9a6c05