bigslice: fix infinite loop when buffering flatmap output
Summary:
A previous change subtly changed the definition of Frame.IsValid,
which was used to check for the EOF condition in bigslice.Flatmap.
This would cause the evaluator to hang indefinitely if Flatmap output
was buffered. Fix this by (1) changing Frame.IsValid to Frame.IsZero,
which spells out its intention more clearly; (2) checking the output
buffer's length explicitly.

Reviewers: afields, ysaito, smahadevan

Reviewed By: afields, ysaito, smahadevan

Differential Revision: https://phabricator.grailbio.com/D18756

fbshipit-source-id: e4592b5