bigslice/sliceio: zero values before decoding them with Gob
Summary:
Gob decoding semantics differ based on values! Gob does not set zero
values, presumably as an implementation side effect of not
transmitting those values.

Thus the only safe way to decode using Gob is to decode into zeroed
values, even when the sender and receiver types are guaranteed to
match, as is the case in Bigslice.

This caused values in Bigslice (or fields in structs) to retain
non-zero values across decodes: so, for example, a record with a zero
field may have presented as nonzero if it was decoded into reused
record that had nonzero values.

Reviewers: ysaito, afields

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D20198

fbshipit-source-id: a44ee90