bigslice: always zero memory that's passed to a user
Summary:
We should be extra careful to zero memory that's provided to the
user, since any other semantics is really unintuitive and dangerous,
as demonstrated by Gob.

This change simplifies and generalizes the zeroing code from the Gob
change and zeroes all frames that may be passed to users. Notably,
instead of building zeroer functions that must be applied correctly,
the API to the internal zero package caches zeroers based on type.
This results in no measurable performance impact but provides a much
nicer API.

Reviewers: ysaito, afields

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D20286

fbshipit-source-id: d41242b