bigslice/codec, frame: custom type-specific codecs
Summary:
This change adds the ability to register frame ops for custom
encoders and decoders for types that wish to override the default gob
encoding. Encoding and decoding are invoked over vectors of values so
that this may be taken advantage of (which is an advantage not
afforded by simply implementing gob.Gob{Encoder,Decoder}).

We take this opportunity to simplify some of the codec code generally
and provide some performance enhancements for the common case where
we are decoding into frames that are have the same size as the
encoded frames, thus avoiding additional buffering and copying.

Reviewers: pgopal, ysaito, smahadevan, ayip

Reviewed By: ysaito, ayip

Subscribers: joshnewman, ayip

Differential Revision: https://phabricator.grailbio.com/D21306

fbshipit-source-id: 3ea2880