bio/encoding/bamslice: efficient serialization
Summary:
Biogo SAM records are not gob-serializable, and even if they were,
they would be serialized inefficiently. This change implements a
custom codec for Slice<*sam.Record> that is efficient. In particular,
SAM records are columnarized (for improved compression) and
references are deduplicated: each encoding and decoding session
maintains deduplication state.

The change also alters the Bigslice codec interface to better
suit what applications need to do in practice.

Reviewers: ysaito, joshnewman, ayip

Reviewed By: ysaito

Subscribers: joshnewman

Differential Revision: https://phabricator.grailbio.com/D21718

fbshipit-source-id: 5c8371d