bigslice: provide result caching
Summary:
This change implements a new slice combinator, bigslice.Cache, which
caches an underlying slice to a user-defined prefix. On subsequent
invocations, the computation is shortcutted and results are read
directly from the cached output. Bigslice does not provide automatic
cache coherency as in Reflow; the user must guarantee that the
computation is correct in the presence of caching. (Of course they
may implement code in Go to perform such checks automatically.)

Reviewers: joshnewman, ysaito

Reviewed By: joshnewman

Subscribers: ayip, afields

Differential Revision: https://phabricator.grailbio.com/D22494

fbshipit-source-id: 9bc0dfd