bigslice: implement user-defined prefixes
Summary:
This change endows the type of slice with a prefix. Prefixes (which
default to 1) determine the number of columns in the slice or frame's
prefix that determine the key of that slice or frame for operations
that require keys. For frames, the prefix determines which columns
are hashed and compared; for slices, the prefix determines which
columns are used as keys for reduce operations.

Reviewers: jschellenberger, ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D24078

fbshipit-source-id: 8308dc5