bigslice/exec: fix reuse of results for tasks that require (re-)shuffling
Summary:
The slice compiler blindly modify tasks even when they were not part
of the compilation session. This broke reuse because the compiler
then assumed it could get inlined partitioning (and combining) from
the set of dependency tasks. In the case of result reuse, this is not
the case: we have already computed the un-partitioned (i.e.,
un-shuffled) tasks. This change fixes this by detecting reuse and, if
needed, inserting a set of tasks for (re-)shuffling the output of a
previous run.

Reviewers: jschellenberger, jcharumilind, pgopal

Reviewed By: pgopal

Subscribers: joshnewman

Differential Revision: https://phabricator.grailbio.com/D27854

fbshipit-source-id: 540391d