bigslice: simplify handling of zeroing memory vis-a-vis sliceio.Reader.
Summary:
Most of bigslice operates on fixed frame.Frame buffers. They are
reused across reads. This historically caused issues in two ways:

- When decode frames with gob, gob could reuse memory already
allocated   in the frames. This was especially problematic when
values were persisted,   such as in the hash tables used by
combiners). - When we passed memory (to populate) to the user, who
might themselves   use gob, or fail to zero unused values.

A number of changes tackled the various issues caused by these
semantics, but it ended up more complex than it had to be; we failed
to see the forest for the trees.

This change attempts to remedy this by putting the onous on reuse
avoidance on sliceio.Reader. We now guarantee the invariant that
sliceio.Reader will never reuse any portion of the frame that is
passed to it. This allows us to remove zeroing code from many points
in the middle of the stack, and relegate them to the the relatively
fewer edges: namely where we invoke the gob decoder, and the (only)
place we call into user code to populate a slice owned by a frame
that's passed into a Reader.

I audited every implementation of sliceio.Reader, and found that this
guarantee was broken nowhere. What's lost to history is whether some
of the zeroing performed by bigslice was due to relieving garbage
collector pressure. If this were the case, and if we need to
reintroduce zeroing, we'll at least know it's for a different
purpose. I also propose that, if such a time comes, we provide a
variant of Frame.Zero(), Frame.Release() that is intended not to zero
memory but to release it. The difference would be that we can skip
doing any work entirely if we are dealing with a value-typed slice.

While we're here, the change also bolsters the tests for the internal
zero package somewhat.

Reviewers: pgopal, jcharumilind

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D27973

fbshipit-source-id: 7a1353a