bigslice: Add `WriterFunc` slice type
Summary:
It is sometimes useful to be able to do a side-effecting computation, similar to what `Scan` enables, while passing through the slice unadulterated; `Scan` results in a unit slice.  This revision implements a `WriterFunc` slice type.

The implementation passes the slice data to a provided `write` function, with a facility for maintaining state between calls.  It is designed as a dual to `ReaderFunc`.

Test Plan: Unit tests; manual testing using `urls` example program modified with intermediate `WriterFunc` slices

Reviewers: marius, pgopal, smahadevan

Reviewed By: marius

Subscribers: afields, joshnewman, jschellenberger

Maniphest Tasks: T18070

Differential Revision: https://phabricator.grailbio.com/D27591

fbshipit-source-id: 4318085