bigslice: add default chunk size flags
Summary:
Temporary solution to the problem we discussed earlier. This can be completely
reverted when adaptive chunk sizing is implemented.

Test Plan:
Existing unit tests.

I also ran bigslicetest, though only a couple of times, rather than in a
rigorous benchmark. With that chunk of salt, it finished in ~80 seconds with
default chunk size 1024 and 32. With the latter it appeared to use less memory,
~40 GiB vs ~30 GiB over two m5.4xlarge, but that's even less meaningful because
there was no GC pressure, etc. Basically, the only conclusion is it seems to
work well enough for temporary relief.

Reviewers: jcharumilind, marius, pgopal

Reviewed By: jcharumilind, marius

Subscribers: afields, jschellenberger

Differential Revision: https://phabricator.grailbio.com/D29087

fbshipit-source-id: 8f03891