bigslice: add securitygroup flag for ec2 system
Summary:
I would like to use bigslice with an instance profile for an IAM
role that is under the 'reference' AWS account. The security group is
defaulting to 'sg-7390e50c', an eng security-group, so I am unable to
run as the security-group does not exist in reference's default VPC.
TidyData needs to write to a bucket under 'reference' and would like to
use bigslice.

If there is an easier way to use bigslice with creds/instance profiles
for an account other than eng, I can abandon this diff.

Test Plan: flags_test.go, test using bigslice flags

Reviewers: ysaito, jcharumilind, pgopal, marius

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D29355

fbshipit-source-id: 83616f4