bigslice: make machine combine buffers optional
Summary:
Machine combine buffers dramatically complicate error handling, and
failure handling was (intentionally) not implemented. I believe also
for most current workloads, the benefit of this optimization is not
dramatic enough to really justify the complexity.

This change reverts the default to task-level buffers instead, so
that they are recoverable by the same mechanism as any other task. A
new session option is provided to turn the optimization on.

Reviewers: jcharumilind, joshnewman, jschellenberger, pgopal

Reviewed By: jcharumilind

Subscribers: joshnewman

Differential Revision: https://phabricator.grailbio.com/D29775

fbshipit-source-id: 508f2cf