base/backgroundcontext: add Get method to get a background context; use in Bigslice
Summary:
This change adds backgroundcontext.Get(). This can be used to in
place of "context.Background()" in libraries that need background
contexts.

The idea is that libraries may or may not be used in the context of a
v23 application, and thus need a way to get a generic background
context that can serve either kind of application.

Reviewers: ysaito, jcharumilind, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D31170

fbshipit-source-id: b33ddee