bigslice: task-exclusive machine access
Summary:
It is often useful to reserve the entirety of a machine for a single
task; for example, an individual map may utilize internal parallelism
because it is inherently more efficient to do so (or it is readily
available, and refactoring it into Bigslice's model of parallelism
would be cumbersome).

This change implements machine exclusivity by dedicating a machine
manager for such tasks, and arranging for that manager's maximum load
such that only one task can run on the machine at a time. This is
simple, and avoids problems like how to prioritize scheduling
multiple small tasks vs. one big task: in other words, the underlying
assumptions about task granularity remain, but we implement a tool
that makes several granularities available.

The user indicates that a slice task should be given exclusive access
through a new mechanism, Pragmas. Pragmas permit the user to annotate
relevant bigslice operations with the necessary directives. It is
implemented so that it may be extended in the future (e.g., to
enforce a certain machine type for a task, which would have a similar
implementation, parameterizing the machine manager on a particular
system configuration).

Reviewers: jcharumilind, jschellenberger, joshnewman, pgopal

Reviewed By: jcharumilind

Subscribers: sgross, afields

Differential Revision: https://phabricator.grailbio.com/D32400

fbshipit-source-id: 85633d1