bigslice/exec: optimize task location representation'
Summary:
We store task locations in each run request. It was being stored as a
map from task name to machine, both of which contain a lot of
redundant information: machines are repeated, and task operations are
similarly redundant.

We take advantage of the fact that the task graph is identical on all
machines, and index locations by their dependency order. Machine
addresses are de-duplicated by indexing into a simple slice.

Reviewers: ysaito, jcharumilind, pgopal

Reviewed By: ysaito, jcharumilind

Differential Revision: https://phabricator.grailbio.com/D32618

fbshipit-source-id: 18003e7