bigslice: adjust maxprocs to match when machines are exclusively allocated
Summary:
Bigslice limits resources by specifying the maximum amount of
allowable parallelism, and matching machine resources to this. The
ratio between CPUs and procs (i.e., allocatable to tasks) is governed
by the load factor. Bigslice.Exclusive also makes use of the load
factor in order to ensure exclusive allocation on a per-machine
basis.

When using load factor, the maximum parallelism parameter is still
applied to the number of allocated procs, not the underlying CPUs.
However, in exclusive mode, this makes less sense: it is expected
that user-provided code makes use of internal parallelism in order to
speed up computation, and thus we should account for this when
applying these limits.

This change adjusts the parallelism limit for machine managers that
use exclusive allocation to account for the internal parallelism
implied by bigslice.Exclusive.

Bonus: we perform batch allocation of machines from the underlying
system, which, in the case of EC2, should reduce (by tenfold) the
number of API calls made by bigslice jobs.

Reviewers: jcharumilind, pgopal

Reviewed By: jcharumilind

Subscribers: joshnewman, jschellenberger, sgross

Differential Revision: https://phabricator.grailbio.com/D32785

fbshipit-source-id: f6f14da