bigslice/typecheck: location attribution
Summary:
typecheck.Location is a function that can be used to change the
default typecheck failure attributions. It is intended to be used in
a defer statement. We use it in session.Run to to make sure that
function application typechecking errors are attributed to the
callsite, and not inside of the session code.

Reviewers: jcharumilind, ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D32987

fbshipit-source-id: 5d6b3ed