bigslice: Verify common funcs across workers
Summary: `bigslice` depends on the invariant that `Func` definitions are common across machines, which we achieve by prescribing that `Func` must be called in a deterministic order *before* `bigmachine` takes over.  This change adds a step that (imprecisely) verifies it by checking that the locations of the definitions of `Func`s are equivalent across workers.

Test Plan: Add `badfuncs` test program.

Reviewers: pgopal, marius

Reviewed By: marius

Subscribers: rcalef, dfilippova

Differential Revision: https://phabricator.grailbio.com/D33331

fbshipit-source-id: 1ac637b