bigslice: fix recovery of tasks experiencing network errors
Summary:
Bigslice aggressively retries transient network errors (via
bigmachine's RetryCall mechanism). When a task reads output from a
machine which times out (before being marked as lost by the failure
detector), the RPC call returns an errors.Net, which is then passed
over the network to the coordinator, which will simply retry the
call, without resetting the task state (since it is assumed the
errors.Net originated locally, and not by the task).

We fix this by modifying bigmachine's RPC stack to guarantee that
errors.Net errors must originate locally; they never cross network
boundaries.

This change also fixes a second issue which would cause a repeat task
to fail if it was scheduled on the same machine.

A unit test (chaos test?) exercising this condition is included.

Note that error interpretation is still too complicated. We should
rethink how to simplify this.

Reviewers: jcharumilind, jschellenberger, pgopal

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D33460

fbshipit-source-id: 69c0f9a