bigslice/slicemachine: fix lost task marking, adjust chaos testing parameters
Summary:
Tasks were not being properly assigned to slicemachines, and thus
completed tasks were also not properly being marked as LOST when
a machine failed. This was discovered via chaos testing.

While we're at it, also adjust the chaos testing parameters to be a
little more reasonable. They were too aggressive to recover from
in the expected timeframes.

Reviewers: jcharumilind, pgopal

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D33660

fbshipit-source-id: 1e9c447