bigslice/sliceio: Fix scratch buffer resizing
Summary:
Make `(*decodingReader).Read` properly allocate the scratch buffer into which to decode.  `(Frame).Ensure` does not mutate the frame.  It returns the frame that ensures `(Frame).Slice(0, n)`.

Without this fix, the added test will panic with the message "gob reallocated a slice", as the buffer passed to `gob` will not be large enough to decode the frame, and `gob` will allocate a slice large enough to do so.

Test Plan: Add unit test.

Reviewers: pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D33852

fbshipit-source-id: 58d3c09