bigslice/exec: Recover from Unavailable error on Worker.Run
Summary:
Make `bigslice` attempt to recover when the `Worker.Run` RPC call returns an `errors.Fatal` `errors.Unavailable` error.  In practice, this is usually caused by the dependee task's machines unavailability, e.g. task A on machine 1 depends on task B on machine 2, but machine 2 is stopped/unavailable.  This results in `TaskErr`, which results in an error return from `(*Session).Run()`, despite the fact that `bigslice` can likely recover by bringing up a machine to recompute task B.  This revision effects a change to treat `errors.Unavailable` errors as recoverable.

Note that this is imprecise, as it's possible for user code to produce an unrecoverable `Unavailable` error.  However, in practice, I have not seen code that does this.  We should eventually make this precise, but for now, this should help a lot more than it hurts.

Also note that `bigslice` is currently expected retry indefinitely in the case of machine failure, which means that if machines reliably die because of OOMs, for example, it will retry forever.

Test Plan: Manual tests; continues to pass chaos monkey tests

Reviewers: pgopal, marius, joshnewman

Reviewed By: joshnewman

Subscribers: joshnewman, ysaito, jschellenberger

Differential Revision: https://phabricator.grailbio.com/D34304

fbshipit-source-id: 748f6d0