bigslice: Account for failed machine starts
Summary:
Deduct procs from `pending` when we try to start machines and fail.  By not deducting them, the state indicates that there are pending procs/machines, but there were not.  In practice, we hit this state when machines systematically failed to successfully start.  Machines would fail to start, and the state machine would come to a halt, believing that there were machines trying to start.  This just looks like a hang from the user's perspective.

This also simplifies the code a bit, unifying handling of success and errors.

Since this change will allow the state machine to indefinitely attempt to start new machines, add a warning message based on the heuristic of consecutive failed attempts to start machines.

Test Plan: Manual testing

Reviewers: pgopal, marius, joshnewman

Reviewed By: joshnewman

Subscribers: dfilippova, rcalef

Differential Revision: https://phabricator.grailbio.com/D34399

fbshipit-source-id: 7c2832c