bigslice/exec: Add slice-level status
Summary:
Add slice-level status to `/debug/status`.  We currently display task-level status, but it's also nice to know how your slices are progressing, since that's the level of abstraction at which you specify the computation in the first place.

When slices are pipelined, the tasks of the pipeline are considered to be shared between all of the pipelined slices.

It looks like this:
{F924225}

Test Plan: Manual testing; unit test added

Reviewers: pgopal, marius

Reviewed By: marius

Subscribers: joshnewman

Differential Revision: https://phabricator.grailbio.com/D34729

fbshipit-source-id: bff11e9