bigslice/cogroup: properly heapify frame buffers before merging
Summary:
The implementation of Cogroup did not properly initialize the heap;
thus correctness depended on the order the underlying readers were
added. In particular, if the first slice added did not have the lowest key
in the union of keys, then the row for that key would be emitted twice.

Reviewers: jcharumilind, jschellenberger, pgopal

Reviewed By: jcharumilind

Subscribers: afields

Differential Revision: https://phabricator.grailbio.com/D36003

fbshipit-source-id: 7aeac54