bigslice: add Materialize pragma which interrupts pipelines
Summary:
Add ExperimentalMaterialize pragma that hints to compilation that pipelines should be interrupted so that intermediate results will be materialized. This may be useful if you want to reuse slice/task results within an invocation that would normally have been pipelined, e.g. you have an expensive operation in the middle of a pipeline. This pragma will cause the pipeline to be interrupted at the slice to which this pragma was applied.

It is marked "experimental" because we are considering other ways of achieving this.

Test Plan: Add unit tests and run them; Patched and testing with classifier evaluation

Reviewers: pgopal, marius

Reviewed By: marius

Subscribers: jschellenberger

Differential Revision: https://phabricator.grailbio.com/D36349

fbshipit-source-id: ed7afab