bigslice/exec: reset retry counter on successful worker reads
Summary:
Reset the retry counter on successful worker reads. The retry counter is incremented when we encounter task data read failures from workers. It was however not being reset upon successful reads, so it would take longer to retry and recover from subsequent failures.

Note that this is effectively a one-line change along with refactoring to make it testable.

Test Plan: Add unit test; test manually

Reviewers: pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D36932

fbshipit-source-id: b9dcf5c